# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Resident Management Service API
  description: Canonical target-state API documentation for Resident Management Service in Project A (Group 1 - Apartment Identity and Resident Management).
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: http://localhost:8000
    description: API Gateway route
paths:
  /api/v1/residents:
    get:
      tags:
        - Residents
      summary: "RES-001: List resident profiles"
      description: Retrieve paginated resident profiles. Requires User JWT with role SYSTEM_ADMINISTRATOR or APARTMENT_MANAGER.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (0-indexed, default 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Page size (1-100, default 20)
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: Filter by profile status (ACTIVE, INACTIVE, ARCHIVED)
          required: false
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, ARCHIVED]
        - name: search
          in: query
          description: Search term across firstName, lastName, or email
          required: false
          schema:
            type: string
        - name: userId
          in: query
          description: Filter by Identity Access user ID (UUID)
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Residents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResidentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags:
        - Residents
      summary: "RES-002: Create resident profile"
      description: Creates a new resident profile. Validates user existence with identity-access-service via internal API.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Resident profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleResidentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: User not found in Identity Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Resident profile already exists for this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '503':
          description: Identity service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/residents/{residentId}:
    get:
      tags:
        - Residents
      summary: "RES-003: Get resident profile"
      description: Retrieve a specific resident profile. Allowed for SYSTEM_ADMINISTRATOR, APARTMENT_MANAGER, or the resident user themselves.
      security:
        - BearerAuth: []
      parameters:
        - name: residentId
          in: path
          required: true
          description: Resident profile UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resident retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleResidentResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Resident not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    patch:
      tags:
        - Residents
      summary: "RES-004: Update resident profile"
      description: Update resident profile fields. Allowed for SYSTEM_ADMINISTRATOR, APARTMENT_MANAGER, or self-service update.
      security:
        - BearerAuth: []
      parameters:
        - name: residentId
          in: path
          required: true
          description: Resident profile UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResidentRequest'
      responses:
        '200':
          description: Resident profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleResidentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Resident not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/internal/residents/{residentId}/validate:
    get:
      tags:
        - Internal
      summary: "RES-INT-001: Validate resident profile"
      description: Validates that a resident profile exists and is active. Accessible only by registered Project A backend services using Service JWT via API Gateway.
      security:
        - BearerAuth: []
      parameters:
        - name: residentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resident validation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentValidationApiResponse'
        '401':
          description: Invalid service token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Caller service not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Resident not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/internal/users/{userId}/relationships:
    get:
      tags:
        - Internal
      summary: "RES-INT-002: Get user relationships"
      description: Retrieves all resident-domain relationships owned by Resident Management for a user. Accessible by registered backend services using Service JWT via API Gateway.
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User relationships retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRelationshipsApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Invalid service token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Caller service not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT token issued/resigned by API Gateway. User JWT for public endpoints, Service JWT for internal endpoints.
  schemas:
    ProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
    CreateProfileRequest:
      type: object
      required:
        - userId
        - firstName
        - lastName
        - email
      properties:
        userId:
          type: string
          format: uuid
          description: Identity Access user UUID
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    UpdateResidentRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    ResidentValidationResponse:
      type: object
      properties:
        residentId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        exists:
          type: boolean
        active:
          type: boolean
    RelationshipItem:
      type: object
      properties:
        profileId:
          type: string
          format: uuid
        profileType:
          type: string
          enum: [RESIDENT, OWNER, TENANT, STAFF]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
    UserRelationshipsData:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        relationships:
          type: array
          items:
            $ref: '#/components/schemas/RelationshipItem'
    PagedData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProfileResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    SingleResidentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProfileResponse'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    PagedResidentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/PagedData'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    ResidentValidationApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/ResidentValidationResponse'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    UserRelationshipsApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserRelationshipsData'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        details:
          nullable: true
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid